+++ R/tabula-internal.R
@@ -6,34 +6,6 @@ is_incidence <- function(x) {
   all(x == 0 | x == 1, na.rm = TRUE)
 }
 
-#' Colour Ramp
-#'
-#' Provides a colour scheme that map a `numeric` vector to colours.
-#' @param x A [`numeric`] vector to be mapped to colours.
-#' @param palette A vector of colors.
-#' @param from A length-two [`numeric`] vector specifying the input range.
-#'  If `NULL`, is calculated from the range of `x`.
-#' @param midpoint A [`numeric`] value specifying the midpoint of `x`. If not
-#'  `NULL`, `x` will be rescaled to have the specified `midpoint`.
-#' @return A [`character`] vector of colors.
-#' @keywords internal
-#' @noRd
-color_ramp <- function(x, palette, from = range(x, na.rm = TRUE),
-                       midpoint = NULL) {
-  ## Rescale to 0-1
-  to <- c(0, 1)
-  if (!is.null(midpoint) && is.numeric(midpoint)) {
-    extent <- 2 * max(abs(from - midpoint))
-    z <- (x - midpoint) / extent * diff(to) + mean(to)
-  } else {
-    z <- scale_range(x, to = to, from = from)
-  }
-
-  col <- grDevices::colorRamp(palette)(z)
-  col <- grDevices::rgb(col, maxColorValue = 255)
-  col
-}
-
 #' Rolling Sum
 #'
 #' @param x A [`numeric`] vector.
@@ -43,4 +15,21 @@ color_ramp <- function(x, palette, from = range(x, na.rm = TRUE),
 #' @noRd
 roll_sum <- function(x, n = 2) {
   utils::tail(cumsum(x) - cumsum(c(rep(0, n), utils::head(x, -n))), -n + 1)
+}
+
+#' Plotting Dimensions of Character Strings
+#'
+#' Convert string length in inch to number of (margin) lines.
+#' @param x A [`character`] vector of string whose length is to be calculated.
+#' @param ... Further parameter to be passed to [graphics::strwidth()]`, such as
+#'  `cex`.
+#' @return
+#'  A [`numeric`] vector (maximum string width in units of margin lines).
+#' @note For internal use only.
+#' @family graphic tools
+#' @keywords internal
+#' @noRd
+inch2line <- function(x, ...) {
+  (max(graphics::strwidth(x, units = "inch", ...)) /
+     graphics::par("cin")[2] + graphics::par("mgp")[2]) * graphics::par("cex")
 }